+++ rollup.config.ts
@@ -1,8 +1,9 @@
 import commonjs from "@rollup/plugin-commonjs";
-import nodeResolve from "@rollup/plugin-node-resolve";
+import nodeResolver from "@rollup/plugin-node-resolve";
 import external from "rollup-plugin-peer-deps-external";
 import terser from "@rollup/plugin-terser";
 import typescript from "rollup-plugin-typescript2";
+import { cleanDist } from "./rollup.plugin";
 
 const input = "src/index.ts";
 
@@ -17,13 +18,14 @@ const plugins = [
   commonjs({
     include: /\/node_modules\//,
   }),
-  nodeResolve(),
+  nodeResolver(),
   terser({
     output: { comments: false },
     compress: {
       drop_console: true,
     },
   }),
+  cleanDist,
 ];
 
 export default [
@@ -35,7 +37,6 @@ export default [
       name: "ReactUtilities",
       sourcemap: true,
       exports: "named",
-      sourcemap: true,
       interop: "auto",
     },
     plugins,
@@ -48,7 +49,6 @@ export default [
       name: "ReactUtilities",
       sourcemap: true,
       exports: "named",
-      sourcemap: true,
     },
     plugins,
   },
@@ -60,7 +60,6 @@ export default [
       name: "ReactUtilities",
       sourcemap: true,
       exports: "named",
-      sourcemap: true,
     },
     plugins,
   },