+++ src/lib/hooks/useConstructor.ts
@@ -1,15 +1,16 @@
-import { useRef } from "react";
+import { useState } from "react";
 
 /**
  * Custom hook that runs the provided callback only once during the component's lifecycle.
  * @param callback - Function to be executed once on component mount.
  */
 const useConstructor = (callback: () => void) => {
-  const calledRef = useRef(false);
-  if (!calledRef.current) {
-    callback();
-    calledRef.current = true;
+  const [hasBeenCalled, setHasBeenCalled] = useState(false);
+  if (hasBeenCalled) {
+    return;
   }
+  callback();
+  setHasBeenCalled(true);
 };
 
 export { useConstructor };